Added simple thumb TC. Updated doc
Former-commit-id: 2c83a55f8aabd21192ba88d0e830288b454cc1d5

+++ src/scripts/run_sim.pl
@@ -29,6 +29,7 @@ my $DUMP_SIZE                   = $Config{'DUMP_SIZE'};
 my $MAX_CLOCK_CYCLES            = $Config{'MAX_CLOCK_CYCLES'};
 my $TLB_DEBUG                   = $Config{'DEFINE_TLB_DEBUG'};
 my $STALL                       = $Config{'ALLOW_STALLS'};
+my $TX_TERM                     = $Config{'UART_TX_TERMINAL'};
 
 # System configuration.
 my $DATA_CACHE_SIZE             = $Config{'DATA_CACHE_SIZE'};
@@ -92,7 +93,13 @@ my $IVL_OPTIONS .=
    $IVL_OPTIONS .= " $ZAP_HOME/src/rtl/*/*.v $ZAP_HOME/src/testbench/cpu/*.v -o $VVP_PATH -gstrict-ca-eval -Wall -g2001 -Winfloop -DSEED=$SEED -DMEMORY_IMAGE=\\\"$PROG_PATH\\\" ";
 
 $IVL_OPTIONS .= " -DVCD_FILE_PATH=\\\"$VCD_PATH\\\" "; 
-$IVL_OPTIONS .= " -DUART_FILE_PATH=\\\"$UART_PATH\\\" "; 
+
+if ( $TX_TERM) { 
+        $IVL_OPTIONS .= " -DUART_FILE_PATH=\\\"$UART_PATH\\\" "; 
+} else {
+        $IVL_OPTIONS .= " -DUART_FILE_PATH=\\\"/dev/null\\\" ";
+}
+
 $IVL_OPTIONS .= " -Pzap_test.RAM_SIZE=$RAM_SIZE -Pzap_test.START=$DUMP_START -Pzap_test.COUNT=$DUMP_SIZE -DLINUX -Pzap_test.STORE_BUFFER_DEPTH=$SBUF_DEPTH ";
 $IVL_OPTIONS .= " -Pzap_test.BP_ENTRIES=$BP -Pzap_test.FIFO_DEPTH=$FIFO ";
 $IVL_OPTIONS .= " -Pzap_test.DATA_SECTION_TLB_ENTRIES=$DATA_SECTION_TLB_ENTRIES ";
@@ -134,15 +141,19 @@ print HH '$display("Simulation Complete. All checks (if any) passed.");$finish;'
 print "*I: Rand is $SEED...\n";
 print "iverilog $IVL_OPTIONS\n";
 
-system("rm -f $UART_PATH");    # Remove UART file.
-system("mknod $UART_PATH p");  # Create a UART output FIFO file.
+if ( $TX_TERM ) {
+        system("rm -f $UART_PATH");    # Remove UART file.
+        system("mknod $UART_PATH p");  # Create a UART output FIFO file.
+}
 
 # UART output monitor.
 die "Error: XTerm could not be found!" if system("which xterm");
 die "Error: Icarus Verilog could not be found!" if system("which iverilog");
 
-print "Setting up UART output monitor";
-system("xterm -T 'TB UART Output' -hold -e 'cat $UART_PATH ; echo ; echo ------------------ ; echo UART_Output_Complete ; echo ------------------' &");
+if ( $TX_TERM == 1 ) {
+        print "Setting up UART output monitor\n";
+        system("xterm -T 'TB UART Output' -hold -e 'cat $UART_PATH ; echo ; echo ------------------ ; echo UART_Output_Complete ; echo ------------------' &");
+}
 
 die "*E: Verilog Compilation Failed!\n" if system("iverilog $IVL_OPTIONS");
 die "*E: VVP execution error!\n" if system("vvp $VVP_PATH | tee $LOG_FILE_PATH");
